+++ qupla/consttypeinfo.go
@@ -6,17 +6,17 @@ import (
 )
 
 type ConstTypeFieldInfo struct {
-	Offset int64
-	Size   int64
+	Offset int
+	Size   int
 }
 
 type ConstTypeInfo struct {
 	name   string
-	size   int64
+	size   int
 	Fields map[string]*ConstTypeFieldInfo
 }
 
-func NewConstTypeInfo(name string, size int64) *ConstTypeInfo {
+func NewConstTypeInfo(name string, size int) *ConstTypeInfo {
 	return &ConstTypeInfo{
 		name:   name,
 		size:   size,
@@ -24,7 +24,7 @@ func NewConstTypeInfo(name string, size int64) *ConstTypeInfo {
 	}
 }
 
-func (e *ConstTypeInfo) GetConstValue() int64 {
+func (e *ConstTypeInfo) GetConstValue() int {
 	return e.size
 }
 
@@ -36,7 +36,7 @@ func (e *ConstTypeInfo) GetSource() string {
 	return ""
 }
 
-func (e *ConstTypeInfo) Size() int64 {
+func (e *ConstTypeInfo) Size() int {
 	return 0
 }
