+++ qupla/lutdef.go
@@ -16,8 +16,8 @@ type LutDef struct {
 //	LutDef.Name = Name
 //}
 
-func (lutDef *LutDef) Size() int64 {
-	return int64(lutDef.OutputSize)
+func (lutDef *LutDef) Size() int {
+	return lutDef.OutputSize
 }
 
 func (lutDef *LutDef) Lookup(res, args Trits) bool {
